add unit tests for tasks and categories using mocks

+++ todo-app-backend/app/services/task.py
@@ -4,6 +4,12 @@ from app.repositories.task import TaskRepository
 from app.schemas.task import TaskCreateSchema, TaskSchema, TaskUpdateSchema
 
 
+class TaskNotFoundError(Exception):
+    def __init__(self, task_id: str) -> None:
+        self.task_id = task_id
+        super().__init__(f"Задача с id='{task_id}' не найдена.")
+
+
 class TaskService:
     def __init__(self, db: Session) -> None:
         self.db = db
@@ -20,6 +26,10 @@ class TaskService:
 
     def update_task(self, task_id: str, task_update: TaskUpdateSchema) -> TaskSchema:
         task_for_update = self.task_repository.get_by_id(task_id=task_id)
+
+        if not task_for_update:
+            raise TaskNotFoundError(task_id=task_id)
+
         if task_update.title is not None:
             task_for_update.title = task_update.title
         if task_update.completed is not None:
@@ -30,5 +40,10 @@ class TaskService:
 
     def delete_task(self, task_id: str) -> None:
         task_for_delete = self.task_repository.get_by_id(task_id=task_id)
+
+        # Если задачи нет в базе сразу выбрасываем исключение
+        if not task_for_delete:
+            raise TaskNotFoundError(task_id=task_id)
+
         self.task_repository.delete(task_for_delete)
         self.db.commit()
